refactor: split observer goal phase declare-war scoring cases (#4310)
Extract declare-war scoring groups from the 437-line phase/cases module
into a sibling cases file so the 400–449 densify band gains headroom under
Slice D remainder.

Refs #4310
